Created model and tests for daily weather.
Created the OpenWeatherDailyDataRetriever class that implements the DailyWeatherDataRetriever interface.

Created the tests for the OpenWeatherDailyDataRetriever.

Implemented a APIDataRetriever class that implements a new DataRetreiver class.  The purpose of the class was to allow for better testing/mocking.  The object gets passed into any data retriever.

App for service updated to rebind both objects for current weather and weather daily forecast.

Renamed WeatherDataRetriever to CurrentWeatherDataRetriever.