Created model and tests for Favorited Locations.
Added the WeatherLocation and WeatherLocationSerializer classes to the gui project.  The WeatherLocation class simply wraps weather location data.  The serializer class handles saving and loading favorited weather locations from file.

Added the tests for the aforementioned classes.

Moved many classes in the shared project around/from the service to the shared project.  This includes: MeasurementUnits from the service project, adding an enums and interfaces folder to the shared project.

Added the getters/setters for measurement units for the data retriever interfaces and added the tests for those.