Added model for searching weather locations
Added the city.list.json file to the project.  This contains the registry of locations that the autocomplete feature checks for weather locations from.

Moved the WeatherLocation class to the shared project and altered imports for it.

Created the WeatherLocationSearcher class for the service and the LocationSearcher on the shared project for RMI use.