Updated Documentation and Setup Script.
Use case diagram updated with the two new Use Cases: Check Hourly Weather and Check Weather Map.  Check Time Period Based Weather was renamed to only Check Daily Weather.

The config script was renamed from 'config.sh' to 'setup.sh' and moved to the root of the repository.  Also added a kill feature to the script which terminates the Java RMI server if it is running.