In progress for creating model for Weather Map.
src/map_site directory created.  This directory contains an html page that uses leaflet and openweathermap javascript libraries to produce a weather radar map.
The plan now is to create a WeatherMapDataRetriever object that sends the website data to the client, so that that the gui app can build the website locally
in a webview/webengine.

Removed an unused import in WeatherLocationSearcherTests.java