Added functionality to display weather map
Added fucntionality to fetch the weather radar data from the service and display it on the weather map gui side.

Added functionality to disable the search bar when loading weather data on the landing page.

Added functionality to disable the settings menu when loading weather data on the landing page.