Created CI-Checks.yml.
This workflow runs when pull-requests are created for the Development branch.  It contains a single job, "merge-feature" that runs on windows-latest that clones the repo, sets up JDK, and then runs verify on each of the three projects within this repo.  It should fail upon not meeting the required 90% code coverage for view and model code.